chore: wire up rollouts
