+++ ui/src/components/flags/FlagForm.tsx
@@ -434,9 +434,9 @@ export default function FlagForm(props: { flag?: IFlag }) {
                     {selectedTab == 'variants' && (
                       <Variants variants={variants!} />
                     )}
-                    {/* {selectedTab == 'rollouts' && (
-                        <Rollouts flag={flag} rollouts={rollouts!} />
-                      )} */}
+                    {selectedTab == 'rollouts' && (
+                      <Rollouts flag={flag} rollouts={rollouts!} />
+                    )}
                     {selectedTab == 'rules' && (
                       <Rules flag={flag} rules={rules!} />
                     )}
