+++ ui/src/components/rollouts/RolloutForm.tsx
@@ -1,10 +1,7 @@
 import { Dialog } from '@headlessui/react';
 import { XMarkIcon } from '@heroicons/react/24/outline';
 import { FieldArray, Form, Formik } from 'formik';
-import { useSelector } from 'react-redux';
-
-import { useUpdateRolloutMutation } from '~/app/flags/rolloutsApi';
-import { selectCurrentNamespace } from '~/app/namespaces/namespacesApi';
+import { useState } from 'react';
 
 import { Button } from '~/components/Button';
 import Loading from '~/components/Loading';
@@ -22,8 +19,6 @@ import {
 } from '~/types/Segment';
 
 import { useError } from '~/data/hooks/error';
-import { useSuccess } from '~/data/hooks/success';
-import { cls } from '~/utils/helpers';
 
 const rolloutRuleTypes = [
   {
@@ -38,102 +33,67 @@ const rolloutRuleTypes = [
   }
 ];
 
-type EditRolloutFormProps = {
+type RolloutFormProps = {
   setOpen: (open: boolean) => void;
   onSuccess: () => void;
-  flagKey: string;
-  rollout: IRollout;
+  createRollout: (rollout: IRollout) => void;
+  rank: number;
   segments: ISegment[];
 };
 
 interface RolloutFormValues {
-  description: string;
+  type: string;
+  description?: string;
+  segmentKeys?: FilterableSegment[];
   operator?: SegmentOperatorType;
   percentage?: number;
-  segmentKeys?: FilterableSegment[];
   value: string;
 }
 
-export default function EditRolloutForm(props: EditRolloutFormProps) {
-  const { setOpen, onSuccess, flagKey, rollout, segments } = props;
+export default function RolloutForm(props: RolloutFormProps) {
+  const { setOpen, onSuccess, rank, segments, createRollout } = props;
 
   const { setError, clearError } = useError();
-  const { setSuccess } = useSuccess();
 
-  const namespace = useSelector(selectCurrentNamespace);
+  const [rolloutRuleType, setRolloutRuleType] = useState(RolloutType.THRESHOLD);
 
-  const segmentOperator =
-    rollout.segment && rollout.segment.segmentOperator
-      ? rollout.segment.segmentOperator
-      : SegmentOperatorType.OR;
-
-  const [updateRollout] = useUpdateRolloutMutation();
   const handleSegmentSubmit = (values: RolloutFormValues) => {
-    let rolloutSegment = rollout;
-    rolloutSegment.threshold = undefined;
-
-    return updateRollout({
-      namespaceKey: namespace.key,
-      flagKey,
-      rolloutId: rollout.id,
-      values: {
-        ...rolloutSegment,
-        description: values.description,
-        segment: {
-          segmentKeys: values.segmentKeys?.map((s) => s.key),
-          segmentOperator: values.operator,
-          value: values.value === 'true'
-        }
+    createRollout({
+      rank,
+      type: rolloutRuleType,
+      description: values.description,
+      segment: {
+        segments: values.segmentKeys?.map((s) => s.key),
+        segmentOperator: values.operator,
+        value: values.value === 'true'
       }
-    }).unwrap();
+    });
+    return Promise.resolve();
   };
 
   const handleThresholdSubmit = (values: RolloutFormValues) => {
-    let rolloutThreshold = rollout;
-    rolloutThreshold.segment = undefined;
-
-    return updateRollout({
-      namespaceKey: namespace.key,
-      flagKey,
-      rolloutId: rollout.id,
-      values: {
-        ...rolloutThreshold,
-        description: values.description,
-        threshold: {
-          percentage: values.percentage || 0,
-          value: values.value === 'true'
-        }
+    createRollout({
+      rank,
+      type: rolloutRuleType,
+      description: values.description,
+      threshold: {
+        percentage: values.percentage || 0,
+        value: values.value === 'true'
       }
-    }).unwrap();
+    });
+    return Promise.resolve();
   };
-
-  const initialValue =
-    rollout.type === RolloutType.THRESHOLD
-      ? rollout.threshold?.value
-        ? 'true'
-        : 'false'
-      : rollout.segment?.value
-        ? 'true'
-        : 'false';
 
   return (
     <Formik
       enableReinitialize
       initialValues={{
-        type: rollout.type,
-        description: rollout.description || '',
-        segmentKeys: segments.flatMap((s) =>
-          rollout.segment?.segmentKeys?.includes(s.key)
-            ? {
-                ...s,
-                displayValue: s.name,
-                filterValue: s.key
-              }
-            : []
-        ),
-        operator: segmentOperator,
-        percentage: rollout.threshold?.percentage,
-        value: initialValue
+        type: rolloutRuleType,
+        description: '',
+        segmentKeys: [],
+        operator: SegmentOperatorType.OR,
+        percentage: 50, // TODO: make this 0?
+        value: 'true'
       }}
       validate={(values) => {
         if (values.type === RolloutType.SEGMENT) {
@@ -143,10 +103,7 @@ export default function EditRolloutForm(props: EditRolloutFormProps) {
             };
           }
         } else if (values.type === RolloutType.THRESHOLD) {
-          if (
-            values.percentage &&
-            (values.percentage < 0 || values.percentage > 100)
-          ) {
+          if (values.percentage < 0 || values.percentage > 100) {
             return {
               percentage: true
             };
@@ -156,9 +113,9 @@ export default function EditRolloutForm(props: EditRolloutFormProps) {
       onSubmit={(values, { setSubmitting }) => {
         let handleSubmit = async (_values: RolloutFormValues) => {};
 
-        if (rollout.type === RolloutType.SEGMENT) {
+        if (rolloutRuleType === RolloutType.SEGMENT) {
           handleSubmit = handleSegmentSubmit;
-        } else if (rollout.type === RolloutType.THRESHOLD) {
+        } else if (rolloutRuleType === RolloutType.THRESHOLD) {
           handleSubmit = handleThresholdSubmit;
         }
 
@@ -166,7 +123,6 @@ export default function EditRolloutForm(props: EditRolloutFormProps) {
           .then(() => {
             onSuccess();
             clearError();
-            setSuccess('Successfully updated rollout');
             setOpen(false);
           })
           .catch((err) => {
@@ -184,7 +140,7 @@ export default function EditRolloutForm(props: EditRolloutFormProps) {
               <div className="flex items-start justify-between space-x-3">
                 <div className="space-y-1">
                   <Dialog.Title className="text-lg font-medium text-gray-900">
-                    Edit Rollout
+                    New Rollout
                   </Dialog.Title>
                   <MoreInfo href="https://www.flipt.io/docs/concepts#rollouts">
                     Learn more about rollouts
@@ -228,8 +184,11 @@ export default function EditRolloutForm(props: EditRolloutFormProps) {
                               name="type"
                               type="radio"
                               className="h-4 w-4 border-gray-300 text-violet-400 focus:ring-violet-400"
-                              disabled={true}
-                              checked={rolloutRule.id === rollout.type}
+                              onChange={() => {
+                                setRolloutRuleType(rolloutRule.id);
+                                formik.setFieldValue('type', rolloutRule.id);
+                              }}
+                              checked={rolloutRule.id === rolloutRuleType}
                               value={rolloutRule.id}
                             />
                           </div>
@@ -253,7 +212,7 @@ export default function EditRolloutForm(props: EditRolloutFormProps) {
                   </fieldset>
                 </div>
               </div>
-              {rollout.type === RolloutType.THRESHOLD && (
+              {rolloutRuleType === RolloutType.THRESHOLD && (
                 <div className="space-y-1 px-4 sm:grid sm:grid-cols-3 sm:gap-4 sm:space-y-0 sm:px-6 sm:py-5">
                   <label
                     htmlFor="percentage"
@@ -282,77 +241,89 @@ export default function EditRolloutForm(props: EditRolloutFormProps) {
                   </div>
                 </div>
               )}
-              {rollout.type === RolloutType.SEGMENT && (
+              {rolloutRuleType === RolloutType.SEGMENT && (
                 <div className="space-y-1 px-4 sm:grid sm:grid-cols-3 sm:gap-4 sm:space-y-0 sm:px-6 sm:py-5">
                   <div>
                     <label
-                      htmlFor="segmentKeys"
+                      htmlFor="segmentKey"
                       className="block text-sm font-medium text-gray-900 sm:mt-px sm:pt-2"
                     >
                       Segment
                     </label>
                   </div>
                   <div className="sm:col-span-2">
-                    <div>
-                      <FieldArray
-                        name="segmentKeys"
-                        render={(arrayHelpers) => (
-                          <SegmentsPicker
-                            segments={segments}
-                            segmentAdd={(segment: FilterableSegment) =>
-                              arrayHelpers.push(segment)
-                            }
-                            segmentRemove={(index: number) =>
-                              arrayHelpers.remove(index)
-                            }
-                            segmentReplace={(
-                              index: number,
-                              segment: FilterableSegment
-                            ) => arrayHelpers.replace(index, segment)}
-                            selectedSegments={formik.values.segmentKeys}
-                          />
-                        )}
-                      />
-                    </div>
-                    <div className="mt-6 flex space-x-8">
-                      {formik.values.segmentKeys.length > 1 &&
-                        segmentOperators.map((segmentOperator, index) => (
-                          <div className="flex space-x-2" key={index}>
-                            <div>
-                              <input
-                                id={segmentOperator.id}
-                                name="operator"
-                                type="radio"
-                                className={cls(
-                                  'h-4 w-4 border-gray-300 text-violet-400 focus:ring-violet-400'
-                                )}
-                                onChange={() => {
-                                  formik.setFieldValue(
-                                    'operator',
-                                    segmentOperator.id
-                                  );
-                                }}
-                                checked={
-                                  segmentOperator.id === formik.values.operator
-                                }
-                                value={segmentOperator.id}
-                              />
-                            </div>
-                            <div>
-                              <label
-                                htmlFor={segmentOperator.id}
-                                className="block text-sm text-gray-700"
-                              >
-                                {segmentOperator.name}{' '}
-                                <span className="font-light">
-                                  {segmentOperator.meta}
-                                </span>
-                              </label>
-                            </div>
-                          </div>
-                        ))}
-                    </div>
+                    <FieldArray
+                      name="segmentKeys"
+                      render={(arrayHelpers) => (
+                        <SegmentsPicker
+                          segments={segments}
+                          segmentAdd={(segment: FilterableSegment) =>
+                            arrayHelpers.push(segment)
+                          }
+                          segmentRemove={(index: number) =>
+                            arrayHelpers.remove(index)
+                          }
+                          segmentReplace={(
+                            index: number,
+                            segment: FilterableSegment
+                          ) => arrayHelpers.replace(index, segment)}
+                          selectedSegments={formik.values.segmentKeys}
+                        />
+                      )}
+                    />
                   </div>
+                  {formik.values.segmentKeys.length > 1 && (
+                    <>
+                      <div>
+                        <label
+                          htmlFor="operator"
+                          className="block text-sm font-medium text-gray-900 sm:mt-px sm:pt-2"
+                        >
+                          Operator
+                        </label>
+                      </div>
+                      <div>
+                        <div className="sm:col-span-2">
+                          <div className="w-48 space-y-4">
+                            {segmentOperators.map((segmentOperator, index) => (
+                              <div className="flex space-x-4" key={index}>
+                                <div>
+                                  <input
+                                    id={segmentOperator.id}
+                                    name="operator"
+                                    type="radio"
+                                    className="h-4 w-4 border-gray-300 text-violet-400 focus:ring-violet-400"
+                                    onChange={() => {
+                                      formik.setFieldValue(
+                                        'operator',
+                                        segmentOperator.id
+                                      );
+                                    }}
+                                    checked={
+                                      segmentOperator.id ===
+                                      formik.values.operator
+                                    }
+                                    value={segmentOperator.id}
+                                  />
+                                </div>
+                                <div className="mt-1">
+                                  <label
+                                    htmlFor={segmentOperator.id}
+                                    className="block text-sm text-gray-700"
+                                  >
+                                    {segmentOperator.name}{' '}
+                                    <span className="font-light">
+                                      {segmentOperator.meta}
+                                    </span>
+                                  </label>
+                                </div>
+                              </div>
+                            ))}
+                          </div>
+                        </div>
+                      </div>
+                    </>
+                  )}
                 </div>
               )}
               <div className="space-y-1 px-4 sm:grid sm:grid-cols-3 sm:gap-4 sm:space-y-0 sm:px-6 sm:py-5">
@@ -365,7 +336,6 @@ export default function EditRolloutForm(props: EditRolloutFormProps) {
                 <Select
                   id="value"
                   name="value"
-                  value={formik.values.value}
                   options={[
                     { label: 'True', value: 'true' },
                     { label: 'False', value: 'false' }
@@ -403,7 +373,7 @@ export default function EditRolloutForm(props: EditRolloutFormProps) {
                 className="min-w-[80px]"
                 disabled={!formik.isValid || formik.isSubmitting}
               >
-                {formik.isSubmitting ? <Loading isPrimary /> : 'Update'}
+                {formik.isSubmitting ? <Loading isPrimary /> : 'Create'}
               </Button>
             </div>
           </div>
